// server/interfaces/authRole.ts
enum AuthRole {
  CREATE_USER = 'ROLE_CREATE_USER',
  MAINTAIN_ACCESS_ROLES = 'ROLE_MAINTAIN_ACCESS_ROLES',
  MAINTAIN_ACCESS_ROLES_ADMIN = 'ROLE_MAINTAIN_ACCESS_ROLES_ADMIN',
  MAINTAIN_OAUTH_USERS = 'ROLE_MAINTAIN_OAUTH_USERS',
  AUTH_GROUP_MANAGER = 'ROLE_AUTH_GROUP_MANAGER',
  MANAGE_NOMIS_USER_ACCOUNT = 'ROLE_MANAGE_NOMIS_USER_ACCOUNT',
  OAUTH_ADMIN = 'ROLE_OAUTH_ADMIN',
}

export default AuthRole